- Backstage requires 4 configmaps and this PR implements the logic to
create these configmaps.
- Extracted each config data template into separate files and use the
golang [text/template](https://pkg.go.dev/text/template) to parse and
render the data value for each configmap.
- Also hard coded dynamic plugin values since there are incoming changes
pertaining to the CR. Although, I updated the CRD. That would be in a
separate PR.

+++ api/v1alpha1/orchestrator_types.go
@@ -32,10 +32,6 @@ type OrchestratorSpec struct {
 	ArgoCd               ArgoCD               `json:"argocd,omitempty"`
 }
 
-// reuse from the subscription - check from the api/compare with the subscription object
-// do we want to expose all the spec within the inherent subscription
-// inline embedding to add field in the subscription object
-// ask Moti to confirm - breaking changes
 type Subscription struct {
 	Namespace           string `json:"namespace,omitempty"`
 	Channel             string `json:"channel,omitempty"`
@@ -111,7 +107,7 @@ type PluginDetails struct {
 
 type NotificationConfig struct {
 	Enabled   bool   `json:"enabled,omitempty"`
-	Port      int64  `json:"port,omitempty"`
+	Port      int    `json:"port,omitempty"`
 	Sender    string `json:"sender,omitempty"`
 	Recipient string `json:"replyTo,omitempty"`
 }
